Fix bug, unlink images after removing product

+++ ecommerce_website/php/removeProduct.php
@@ -7,16 +7,27 @@ session_start();
 
 $productId = $_GET["id"];
 $storeId = $_SESSION["user"]->getStoreId();
-if ($stmt = $connection->prepare("DELETE FROM product WHERE id = ? AND store_id = ?")) {
-    $stmt->bind_param("ii", $productId, $storeId);
-    $stmt->execute();
-    $stmt->store_result();
 
-    if ($stmt->affected_rows > 0) {
-        echo "Removed Product!";
-    } else {
-        echo "Already Removed!";
+if ($stmt = $connection->query("SELECT path FROM `image` WHERE product_id = " . intval($productId))) {
+
+    while ($image = $stmt->fetch_assoc()) {
+        unlink($image["path"]);
     }
+
+    if ($stmt = $connection->prepare("DELETE FROM product WHERE id = ? AND store_id = ?")) {
+        $stmt->bind_param("ii", $productId, $storeId);
+        $stmt->execute();
+        $stmt->store_result();
+
+        if ($stmt->affected_rows > 0) {
+            echo "Removed Product!";
+        } else {
+            echo "Already Removed!";
+        }
+    } else {
+        echo $connection->error;
+    }
+
 } else {
     echo $connection->error;
 }
